--- src/channels/index.ts
@@ -1,1 +1,0 @@
-export { default as PostMessage  } from './PostMessage.js'
