+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from './agents/index.js'
 export * from './encrypt/index.js'
 export * from './sign/index.js'
-export * from './forge/index.js'
-export * from './channels/index.js'
+export * from './derive/index.js'
+export * from './channel/index.js'
 export * from './hash/index.js'
